fix build issue (#19874)
Not sure how this passed before... Fix kernel build issue for ring
fabric



### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/14139382236
- [x] metal microbench:
https://github.com/tenstorrent/tt-metal/actions/runs/14139367265
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)
- [ ] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes

+++ tt_metal/fabric/hw/inc/edm_fabric/fabric_edm_packet_transmission.hpp
@@ -131,7 +131,7 @@ FORCE_INLINE void execute_chip_unicast_to_local_chip(
                 if constexpr (enable_ring_support) {
                     auto start_trid = RX_CH_TRID_STARTS[rx_channel_id];
                     auto end_trid = start_trid + NUM_TRANSACTION_IDS;
-                    for (size_t i = start_trid; i < end_trid; i++) {
+                    for (int i = start_trid; i < end_trid; i++) {
                         while (!ncrisc_noc_nonposted_write_with_transaction_id_flushed(
                             tt::tt_fabric::edm_to_local_chip_noc, i));
                     }
